+++ vbox/src/script.c
@@ -259,7 +259,7 @@ int vbox_message_info(ClientData cd, Tcl_Interp *ip, int argc, char *argv[])
 				switch (field)
 				{
 					case 1:
-						printstring(ip->result, "%ld", ntohl(header.time));
+						printstring(ip->result, "%ld", (long)ntohl(header.time));
 						break;
 
 					case 2:
